HTML generated for the Omeka Simple Pages
The Smart Tourism Tools (STTs) section of the Observatory is created using the Simple Pages Omeka plugin. In this section there is a right sidebar menu with all the STT taxonomy categories. The dispose of the taxonomy is created in Menu.py. Aggregate_Menu_with_Info.py is responsible for aggregate the menu to each Simple Pages of this section.